+++ .gitignore
@@ -3,3 +3,4 @@ __pycache__
 .idea
 .env
 db.sqlite3
+toDO
